feat: add CUDA support
A new environment variable `BELLMAN_GPU_FRAMEWORK` is introduced to select
CUDA or OpenCL if Bellperson is compiled with support for both.

Currently only the multiexp calculation is supported by CUDA, the FFT
calculation still runs on OpenCL.

BREAKING CHANGE: `gpu` feature is replaced with `cuda` and `opencl` feature.

The `gpu` feature is now split into two pieces, `opencl` and `cuda`, you can
compile the library with either or both.